Switch to rebar3-based hex.pm publication
There are still problems with the mix-based package on hex.pm (see #33).
Rather than sort through them, it seems more straightforward to just
publish this library as a native Erlang / rebar3 package instead.

This change updates the README.md with instructions on how to do that.
It also removes package.exs, which is no longer needed if we use rebar3
exclusively.